Desktop/comp90024_team_61-main/backend/fission/functions/scenario5/requirements.txt
flask
pandas
numpy
elasticsearch==8.10.1
python-dateutil
